// src/types/questionnaire.ts
export type QuestionnaireQuestionType =
  | 'skip'
  | 'alarm'
  | 'text'
  | 'number'
  | 'choice'
  | 'choice-details'
  | 'image-choice'
  | 'multiple-choice'
  | 'questionnaire-choice'
  | 'height-weight'
  | 'goal-weight'
  | 'weight'
  | 'message'
  | 'amh-message-effect'
  | 'hair-loss-treatment-select'
  | 'hair-loss-treatment-add-on'
  | 'medical-history'
  | 'mental-health-result'
  | 'suicide-disclaimer'
  | 'suicide-alarm'
  | 'blood-pressure'
  | 'add-mental-coaching'
  | 'birth-control-treatment-select'
  | 'birth-control-treatment-select-v2'
  | 'ed-treatment-select'
  | 'ed-dosage-select'
  | 'ed-quantity-select'
  | 'hair-loss-photo'
  | 'hair-loss-photo-two'
  | 'add-weight-coaching'
  | 'add-weight-coaching-access'
  | 'emergency-contraception'
  | 'mental-health-provider-schedule'
  | 'primary-care-provider-schedule'
  | 'mental-health-zealthy-program'
  | 'mental-health-coach-schedule'
  | 'provider-schedule'
  | 'weight-loss-coach-schedule'
  | 'mental-health-addon-payment'
  | 'select-visit-type'
  | 'confirmation'
  | 'delivery-address'
  | 'insurance-information'
  | 'insurance-information-compound-weight-loss-refill'
  | 'lab-or-blood-tests'
  | 'responses-reviewed'
  | 'refill-responses-reviewed'
  | 'recurring-intro'
  | 'weight-loss-eligibility'
  | 'weight-loss-ineligible'
  | 'weight-government-insurance'
  | 'weight-loss-medical'
  | 'weight-loss-preference'
  | 'weight-loss-pay'
  | 'now-what'
  | 'now-what-7380'
  | 'wl-insurance-option'
  | 'weight-loss-treatment'
  | 'weight-loss-bundled-treatment'
  | 'compound-weight-loss-refill-treatment'
  | 'compound-weight-loss-refill-recurring-treatment'
  | 'weight-loss-prior-auth'
  | 'weight-loss-checkout-success'
  | 'weight-loss-checkout-success-esp'
  | 'weight-loss-bill-of-rights'
  | 'weight-checkin'
  | 'weight-graph'
  | 'submit-weight-loss-checkin'
  | 'weight-loss-bundle-reorder'
  | 'bundle-weight-loss-reorder-treatment'
  | 'weight-coaching-agreement'
  | 'analyze-weight-loss-results'
  | 'skin-treatment'
  | 'pharmacy-select'
  | 'photo-identification'
  | 'photo-face'
  | 'identity-verification'
  | 'ilv-start'
  | 'ilv-end'
  | 'async-what-happens-next'
  | 'async-what-happens-next-v2'
  | 'skincare-what-next'
  | 'submit-weight-loss-refill'
  | 'submit-compound-weight-loss-refill'
  | 'non-glp1-final'
  | 'submit-non-glp1-meds-request'
  | 'iterative'
  | 'async-mental-health-treatment'
  | 'working-together'
  | 'async-mental-health-start'
  | 'consultation-selection'
  | 'ssn-verification'
  | 'crosscheck-verification'
  | 'complete-visit'
  | 'side-effects'
  | 'enclomiphene-treatment-select'
  | 'at-home-lab'
  | 'async-results'
  | 'async-mental-health-dosages'
  | 'bundled-plan'
  | 'disqualify-glp1'
  | 'non-bundled-plan'
  | 'live-provider-visit'
  | 'performance-protocol'
  | 'pre-workout-blood-pressure'
  | 'pre-workout-treatment-select'
  | 'ed-treatment-select-v2'
  | 'ed-usage-select'
  | 'hair-loss-message'
  | 'hair-loss-review'
  | 'hair-loss-before-after'
  | 'hair-loss-what-next'
  | 'female-hair-loss-select'
  | 'ed-hl-select'
  | 'ed-hl-treatment-select'
  | 'menopause-treatment-select'
  | 'ed-hl-image-question'
  | 'female-hair-loss-info'
  | 'female-hair-loss-treatment-select'
  | 'weight-goal'
  | 'glp1-explanation'
  | 'pricing-options'
  | 'compound-sem-explanation'
  | 'insurance-plan'
  | 'bundled-choice'
  | 'mobile-download'
  | 'oral-semaglutide-plan'
  | 'honesty-privacy'
  | 'prep-insurance'
  | 'gilead-assistance'
  | 'image'
  | 'transition'
  | 'weight-loss-free-consult-insurance'
  | 'free-consult-selection'
  | 'free-consult-precheckout'
  | 'free-consult-medication-selection'
  | 'free-consult-medication-selection-period'
  | 'sleep-selection';

export type QuestionnaireQuestionAnswerOptions = {
  text: string;
  code?: string;
  next?: string;
  system?: string;
  image?: string;
  color?: { text: string; background: string };
  path?: string;
  subText?: string;
};

export type QuestionFollowUp = {
  showIfResponse: string;
  code: string;
  type: string;
  canvas_linkId: string;
  description: string;
  body?: string;
  input_placeholder?: string;
};

export type Question = {
  styles?: {
    [key: string]: {
      [key: string]: string;
    };
  };
  codingSystem?: string;
  description?: string;
  listItems?: string[];
  unorderedList?: string[];
  disclaimers?: string[];
  alerts?: string[];
  allowToSkip?: boolean;
  banner?: string;
  header: string;
  subheader?: string;
  subheaderBold?: string;
  subType?: string;
  label?: string;
  checkboxText?: string;
  type: QuestionnaireQuestionType;
  canvas_linkId?: string;
  input_placeholder?: string;
  next?: string;
  path?: string;
  followUp?: QuestionFollowUp;
  followUpAlt?: QuestionFollowUp;
  answerOptions?: QuestionnaireQuestionAnswerOptions[];
  buttonText?: string;
  hideHeader?: boolean;
  hideFooter?: boolean;
  footer?: {
    description: string;
    detail?: string;
  };
  dependency?: string; //answers it depends on
  template?: string; //set of questions that needs to be repeated
  required?: boolean;
  index?: number;
  checkItems?: string[];
  transition?: {
    time?: number;
    imageSrc?: string;
    imageSrcMobile?: string;
  };
};

export type QuestionWithName = Question & {
  name: string;
  questionnaire: string;
};

export type QuestionnaireQuestion = {
  [key: string]: Question;
};

export type Questionnaire = {
  skipIntro?: boolean;
  name: QuestionnaireName;
  care?: string;
  version: string;
  canvas_id: string;
  code: string;
  codingSystem: string;
  languages?: {
    en: {
      questions: QuestionnaireQuestion;
      intro?: {
        header: string;
        description?: string;
        listItems?: string[];
        buttonText?: string;
        footnote?: string;
      };
      entry: string;
    };
    esp: {
      questions: QuestionnaireQuestion;
      intro?: {
        header: string;
        description?: string;
        listItems?: string[];
        buttonText?: string;
        footnote?: string;
      };
      entry: string;
    };
  };
  intro?: {
    header: string;
    description?: string;
    listItems?: string[];
    buttonText?: string;
    footnote?: string;
  };
  entry: string;
  questions: QuestionnaireQuestion;
};